requirements.txt
pyopengl
imgui[full]
numpy
pyglm
